Continued merge from upstream, added all the modified JUCE library files.

+++ JuceLibraryCode/modules/juce_core/streams/juce_InputStream.cpp
@@ -175,7 +175,7 @@ String InputStream::readString()
         }
     }
 
-    return String::fromUTF8 (data, (int) i);
+    return String::fromUTF8 (data, i);
 }
 
 String InputStream::readNextLine()
